bottle_emptier: move patch into its own file
