+++ ONI_Together/Patches/World/StructureSyncPatch.cs
@@ -122,14 +122,4 @@ namespace ONI_Together.Patches.World
             __instance.gameObject.AddOrGet<PrintingPodSyncComponent>();
         }
     }
-
-    [HarmonyPatch(typeof(BottleEmptier), nameof(BottleEmptier.OnSpawn))]
-    public static class BottleEmptierSpawnPatch
-    {
-        public static void Postfix(BottleEmptier __instance)
-        {
-            using var _ = Profiler.Scope();
-            __instance.gameObject.AddOrGet<BottleEmptierSyncer>();
-        }
-    }
 }
